mb/emulation/qemu-i440fx: prepare fw_cfg for romstage usage
Add separate functions for selecting the port and reading the port.
Romstage can now read incremental from the data port.

Reviewed-on: https://review.coreboot.org/c/30846
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/mainboard/emulation/qemu-i440fx/fw_cfg.h
@@ -14,7 +14,7 @@
 #define FW_CFG_H
 #include "fw_cfg_if.h"
 
-void fw_cfg_get(int entry, void *dst, int dstlen);
+void fw_cfg_get(uint16_t entry, void *dst, int dstlen);
 int fw_cfg_check_file(FWCfgFile *file, const char *name);
 int fw_cfg_max_cpus(void);
 unsigned long fw_cfg_smbios_tables(int *handle, unsigned long *current);
